# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "workpad",
    "description": "OpenAPI documentation for HTTP applications",
    "version": "0.1.0"
  },
  "externalDocs": {
    "url": "/docs/openapi.json"
  },
  "paths": {
    "/api/notes": {
      "post": {
        "tags": [
          "Notes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotesCreateNoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesNote"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/by-date": {
      "get": {
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "2026-01-01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesNote"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/excerpts": {
      "put": {
        "tags": [
          "Notes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotesUpdateNoteExcerptRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/notes/excerpts/{project}": {
      "get": {
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Project A"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotesNoteExcerpt"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notes/for-month": {
      "get": {
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 2026
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "prefix",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Proj"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectsProject"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectsCreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsProject"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/batch": {
      "post": {
        "tags": [
          "Projects"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectsCreateMultipleProjectsRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectsProject"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{name}": {
      "delete": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Project A"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Project A"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectsProject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiPassLoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLoginJwtResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiAuthErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiPassSignupBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLoginJwtResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiServerErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiAuthErrorResponse": {
        "required": [
          "message",
          "status"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Unauthorized"
          },
          "status": {
            "enum": [
              401
            ],
            "type": "integer"
          }
        }
      },
      "ApiLoginJwtResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        }
      },
      "ApiMeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
      "ApiPassLoginBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "email@site.com"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ApiPassSignupBody": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "email@site.com"
          },
          "password": {
            "minLength": 8,
            "type": "string"
          },
          "password2": {
            "type": "string"
          }
        }
      },
      "ApiServerErrorResponse": {
        "required": [
          "message",
          "status"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Internal Server Error"
          },
          "status": {
            "enum": [
              500
            ],
            "type": "integer"
          }
        }
      },
      "NotesCreateNoteRequest": {
        "required": [
          "html_content"
        ],
        "type": "object",
        "properties": {
          "html_content": {
            "type": "string"
          }
        }
      },
      "NotesExcerptNode": {
        "required": [
          "projects",
          "node"
        ],
        "type": "object",
        "properties": {
          "node": {
            "type": "string",
            "example": "{\"type\":\"paragraph\",\"content\":[{\"type\":\"text\",\"text\":\"Sample excerpt text.\"}]}"
          },
          "projects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Project A",
              "Project B"
            ]
          }
        }
      },
      "NotesNote": {
        "required": [
          "html_content",
          "note_date",
          "id"
        ],
        "type": "object",
        "properties": {
          "html_content": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "note_date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NotesNoteExcerpt": {
        "required": [
          "project_name",
          "excerpt",
          "date",
          "id",
          "note_id"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "note_id": {
            "type": "integer"
          },
          "project_name": {
            "type": "string"
          }
        }
      },
      "NotesUpdateNoteExcerptRequest": {
        "required": [
          "excerpts",
          "date"
        ],
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-01-01"
          },
          "excerpts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotesExcerptNode"
            }
          }
        }
      },
      "ProjectsCreateMultipleProjectsRequest": {
        "required": [
          "projects"
        ],
        "type": "object",
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "example": [
              "Project A",
              " Project B"
            ]
          }
        }
      },
      "ProjectsCreateProjectRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ProjectsProject": {
        "required": [
          "name",
          "created_at"
        ],
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "Bearer"
      }
    }
  }
}